+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oauth2/DummyAuthorizationServiceStrategy.kt
@@ -4,15 +4,17 @@ import at.asitplus.openid.AuthenticationRequestParameters
 import at.asitplus.openid.AuthorizationDetails
 import at.asitplus.openid.OidcUserInfoExtended
 import at.asitplus.openid.OpenIdAuthorizationDetails
+import at.asitplus.wallet.lib.oidvci.OAuth2DataProvider
+import at.asitplus.wallet.lib.openid.DummyOAuth2DataProvider.user
 
 class DummyAuthorizationServiceStrategy(
-    private val user: OidcUserInfoExtended,
     private val scope: String,
 ) : AuthorizationServiceStrategy {
+    @Deprecated("Moved to SimpleAuthorizationService")
     override suspend fun loadUserInfo(
         request: AuthenticationRequestParameters,
         code: String,
-    ): OidcUserInfoExtended? = user
+    ): OidcUserInfoExtended? = null
 
     override fun validScopes(): String = scope
 
@@ -24,4 +26,13 @@ class DummyAuthorizationServiceStrategy(
     override fun filterScope(scope: String): String? = scope
     override fun allCredentialIdentifier(): Collection<String> = listOf()
 
+}
+
+class DummyDataProvider(private val user: OidcUserInfoExtended) : OAuth2DataProvider {
+    override suspend fun loadUserInfo(
+        request: AuthenticationRequestParameters,
+        code: String,
+    ): OidcUserInfoExtended? {
+        return user
+    }
 }